corrected some of the code

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,7 @@
 #Using Puppet, create a manifest that kills a process named killmenow
 exec {'kill':
-commad   => 'pkill -f killmenow',
-provider => shell,
+  command   => 'pkill -f killmenow',
+  path     => '/bin:/usr/bin',
+  user     => 'root',
+  onlyif   => 'pgrep -f killmenow',
 }
